backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
groq==0.9.0
spacy==3.8.0
pydantic==2.5.2
python-multipart==0.0.6
pandas==2.0.3
pydantic-settings
pymongo==4.6.1